Refactor : 서비스 계층 삭제에 따른 코드 수정
Optional 처리 예정
